requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==3.6.2
async-timeout==4.0.2
attrs==23.1.0
blinker==1.6.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.1
frozenlist==1.3.3
future==0.18.3
h11==0.14.0
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
line-bot-sdk==2.4.2
logger==1.4
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.5
pydantic==1.10.7
PyYAML==6.0
requests==2.29.0
sniffio==1.3.0
starlette==0.26.1
tqdm==4.65.0
typing-extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.3.2
yarl==1.9.2
zipp==3.15.0
